Crate problem054

Problem 54 (Poker hands)

Solution Detail

There are no tricks, here; we must simply parse each pair of hands, calculate their poker value and check for a winner.

To ease comparison between two hands, they can be converted into a canonical form - a structure consisting of:

The type of poker hand (Straight flush, Four of a kind, Full house...)

The distinct ranks of cards in the hand, sorted first by count, then by value

Then lexicographic ordering of these structures corresponds to ordering of poker hands.

Structs

Card
A structure holding a single playing card.

PokerHand
A structure representing a poker hand in a canonical form that means two hands can be compared easily.

Enums

PokerHandType
The different classes of poker hand that are available.

Rank
The ranks that a playing car can hold.

Suit
The suits that a playing card can belong to.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the number of games of poker from the given hands that were won by the first player.